+++ models/art.js
@@ -1,5 +1,48 @@
 const { Model, DataTypes } = require('sequelize');
 const bcrypt = require('bcrypt');
 const sequelize = require('../config/connection');
+const { TABLOCK } = require('sequelize/types/table-hints');
 
-class User extends Model
+class Art extends Model {}
+
+Art.init(
+    {
+        id: {
+            type: DataTypes.INTEGER,
+            allowNull: false,
+            primaryKey: true,
+            autoIncrement: true,
+          },
+          name: {
+            type: DataTypes.STRING,
+            allowNull: false,
+          },
+          image: {
+            type: DataTypes.STRING,
+            allowNull: false,
+          },
+          description: {
+            type: DataTypes.STRING,
+          },
+          date_added: {
+            type: DataTypes.DATE,
+            allowNull: false,
+          },
+          artist_key: {
+            type: DataTypes.INTEGER,
+            allowNull: false,
+            references: {
+                model: 'user',
+                key: 'id',
+            }
+          }
+    },
+    {
+        sequelize,
+        freezeTableName: true,
+        underscored: true,
+        modelName: 'art',
+      },
+)
+
+module.exports= Art;
