--- models/artist.js
@@ -1,48 +1,0 @@
-const { Model, DataTypes } = require('sequelize');
-const bcrypt = require('bcrypt');
-const sequelize = require('../config/connection');
-const { TABLOCK } = require('sequelize/types/table-hints');
-
-class Art extends Model {}
-
-Art.init(
-    {
-        id: {
-            type: DataTypes.INTEGER,
-            allowNull: false,
-            primaryKey: true,
-            autoIncrement: true,
-          },
-          name: {
-            type: DataTypes.STRING,
-            allowNull: false,
-          },
-          image: {
-            type: TOBEDETERMINED,
-            allowNull: false,
-          },
-          description: {
-            type: DataTypes.STRING,
-          },
-          date_added: {
-            type: DataTypes.DATE,
-            allowNull: false,
-          },
-          artist_key: {
-            type: DataTypes.INTEGER,
-            allowNull: false,
-            references: {
-                model: 'user',
-                key: 'id',
-            }
-          }
-    },
-    {
-        sequelize,
-        freezeTableName: true,
-        underscored: true,
-        modelName: 'art',
-      },
-)
-
-module.exports= Art;
